requirements.txt
# LLM APIs
openai>=1.0.0
anthropic>=0.20.0
groq>=0.5.0

# HuggingFace ecosystem
transformers>=4.40.0
sentence-transformers>=2.7.0
huggingface-hub>=0.22.0

# PyTorch
torch>=2.2.0

# Prompt handling
langchain>=0.1.0
langchain-openai>=0.1.0

# Dashboard
gradio>=4.0.0

# Config validation
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
numpy>=1.26.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0